[arcane,aleph/petsc] Déplace les définition de 'IAlephPETSc.h' dans 'AlephPETSc.cc'.
Supprime le fichier 'IAlephPETSc.h' qui n'est plus utilisé.
