+++ arcane/src/arcane/aleph/petsc/srcs.cmake
@@ -1,5 +1,4 @@
 set(ARCANE_SOURCES
   AlephPETSc.cc
-  IAlephPETSc.h
   )
